[Refactor] Move DiagnoseDaemon lifecycle to GlobalEnv
