+++ be/src/runtime/exec_env.cpp
@@ -68,7 +68,6 @@
 #include "platform/platform_env.h"
 #include "platform/store_path.h"
 #include "runtime/batch_write/batch_write_mgr.h"
-#include "runtime/diagnose_daemon.h"
 #include "runtime/lookup_stream_mgr.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_filter_cache.h"
@@ -160,7 +159,7 @@ void ExecEnv::_refresh_service_contexts() {
     _runtime_services.spill_dir_mgr = _compute_env == nullptr ? nullptr : _compute_env->spill_dir_mgr();
     _runtime_services.global_spill_manager = _compute_env == nullptr ? nullptr : _compute_env->global_spill_manager();
     _runtime_services.connector_sink_spill_executor = _connector_sink_spill_executor;
-    _runtime_services.diagnose_daemon = _diagnose_daemon;
+    _runtime_services.diagnose_daemon = global_env->diagnose_daemon();
 
     _agent_services.agent_server = _agent_server;
 
@@ -281,9 +280,6 @@ Status ExecEnv::init(const std::vector<StorePath>& store_paths, ProcessMetricsRe
     RETURN_IF_ERROR_WITH_WARN(StorageEnv::GetInstance()->init(storage_env_options), "init StorageEnv failed");
 
     RETURN_IF_ERROR(global_env->init_lake_thread_pools(process_metrics));
-
-    _diagnose_daemon = new DiagnoseDaemon();
-    RETURN_IF_ERROR(_diagnose_daemon->init());
 
     auto capacity = std::max<size_t>(config::query_cache_capacity, 4L * 1024 * 1024);
     RETURN_IF_ERROR(_compute_env->init_query_cache(capacity));
@@ -497,12 +493,6 @@ void ExecEnv::stop() {
         component_times.emplace_back("dictionary_cache_pool", MonotonicMillis() - start);
     }
 
-    if (_diagnose_daemon) {
-        start = MonotonicMillis();
-        _diagnose_daemon->stop();
-        component_times.emplace_back("diagnose_daemon", MonotonicMillis() - start);
-    }
-
     start = MonotonicMillis();
     PythonEnvManager::getInstance().close();
     component_times.emplace_back("PythonEnvManager", MonotonicMillis() - start);
@@ -550,7 +540,6 @@ void ExecEnv::destroy() {
     SAFE_DELETE(_lookup_dispatcher_mgr);
     SAFE_DELETE(_batch_write_mgr);
     StorageEnv::GetInstance()->destroy();
-    SAFE_DELETE(_diagnose_daemon);
     DCHECK(_global_env != nullptr);
     _global_env->destroy_thread_pools();
     _query_execution_services.process_metrics = nullptr;
