+++ be/src/service/service_be/starrocks_be.cpp
@@ -163,7 +163,7 @@ void start_be(const std::vector<StorePath>& paths, bool as_cn) {
     DataWorkflowsEnvOptions data_workflows_env_options;
     data_workflows_env_options.exec_env = exec_env;
     data_workflows_env_options.lake_tablet_manager = StorageEnv::GetInstance()->lake_tablet_manager();
-    data_workflows_env_options.diagnose_daemon = exec_env->diagnose_daemon();
+    data_workflows_env_options.diagnose_daemon = global_env->diagnose_daemon();
     data_workflows_env_options.brpc_stub_cache = platform_env->brpc_stub_cache();
     data_workflows_env_options.metrics = process_metrics_registry->root_registry();
     data_workflows_env_options.table_metrics_mgr = process_metrics_registry->table_metrics_mgr();
