+++ be/test/data_workflows/load/tablet_writer/lake_tablets_channel_test.cpp
@@ -36,7 +36,7 @@
 #include "fs/fs_util.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "platform/platform_env.h"
-#include "runtime/exec_env.h"
+#include "runtime/env/global_env.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_metrics.h"
 #include "serde/protobuf_serde.h"
@@ -69,7 +69,7 @@ public:
         _tablet_manager = std::make_unique<lake::TabletManager>(_location_provider, _update_manager.get(), 1024 * 1024);
 
         _load_channel_mgr =
-                std::make_unique<LoadChannelMgr>(_tablet_manager.get(), ExecEnv::GetInstance()->diagnose_daemon(),
+                std::make_unique<LoadChannelMgr>(_tablet_manager.get(), GlobalEnv::GetInstance()->diagnose_daemon(),
                                                  PlatformEnv::GetInstance()->brpc_stub_cache());
 
         auto metadata = new_tablet_metadata(10086);
@@ -204,7 +204,7 @@ protected:
 
         auto load_mem_tracker = std::make_unique<MemTracker>(-1, "", _mem_tracker.get());
         _load_channel = std::make_shared<LoadChannel>(
-                _load_channel_mgr.get(), _tablet_manager.get(), ExecEnv::GetInstance()->diagnose_daemon(),
+                _load_channel_mgr.get(), _tablet_manager.get(), GlobalEnv::GetInstance()->diagnose_daemon(),
                 PlatformEnv::GetInstance()->brpc_stub_cache(), UniqueId::gen_uid(), next_id(), string(), 1000,
                 std::move(load_mem_tracker));
         TabletsChannelKey key{UniqueId::gen_uid().to_proto(), 0, kIndexId};
