+++ be/test/data_workflows/load/tablet_writer/load_channel_mgr_test.cpp
@@ -22,7 +22,7 @@
 #include "base/utility/defer_op.h"
 #include "common/config_ingest_fwd.h"
 #include "platform/platform_env.h"
-#include "runtime/exec_env.h"
+#include "runtime/env/global_env.h"
 #include "service/brpc_service_test_util.h"
 #include "storage/chunk_helper.h"
 #include "storage/storage_engine.h"
@@ -39,7 +39,7 @@ public:
 protected:
     void SetUp() override {
         _mem_tracker = std::make_unique<MemTracker>(-1);
-        _load_channel_mgr = std::make_unique<LoadChannelMgr>(nullptr, ExecEnv::GetInstance()->diagnose_daemon(),
+        _load_channel_mgr = std::make_unique<LoadChannelMgr>(nullptr, GlobalEnv::GetInstance()->diagnose_daemon(),
                                                              PlatformEnv::GetInstance()->brpc_stub_cache());
         _node_id = 100;
         _db_id = 100;
