+++ be/test/data_workflows/load/tablet_writer/local_tablets_channel_test.cpp
@@ -34,7 +34,7 @@
 #include "gen_cpp/internal_service.pb.h"
 #include "gutil/walltime.h"
 #include "platform/platform_env.h"
-#include "runtime/exec_env.h"
+#include "runtime/env/global_env.h"
 #include "runtime/mem_tracker.h"
 #include "serde/protobuf_serde.h"
 #include "storage/chunk_helper.h"
@@ -71,10 +71,10 @@ protected:
         _root_profile = std::make_unique<RuntimeProfile>("LoadChannel");
         auto load_mem_tracker = std::make_unique<MemTracker>(-1, "", _mem_tracker.get());
         auto* platform_env = PlatformEnv::GetInstance();
-        _load_channel_mgr = std::make_unique<LoadChannelMgr>(nullptr, ExecEnv::GetInstance()->diagnose_daemon(),
+        _load_channel_mgr = std::make_unique<LoadChannelMgr>(nullptr, GlobalEnv::GetInstance()->diagnose_daemon(),
                                                              platform_env->brpc_stub_cache());
         _load_channel = std::make_shared<LoadChannel>(
-                _load_channel_mgr.get(), nullptr, ExecEnv::GetInstance()->diagnose_daemon(),
+                _load_channel_mgr.get(), nullptr, GlobalEnv::GetInstance()->diagnose_daemon(),
                 platform_env->brpc_stub_cache(), _load_id, _txn_id, string(), 1000, std::move(load_mem_tracker));
         _tablets_channel = new_local_tablets_channel(_load_channel.get(), {_load_id, _sink_id, _index_id},
                                                      _load_channel->mem_tracker(), _root_profile.get(),
